Deprecate other values than string for CreditCard
Test for card number `6304100000000008` fails on windows because a number this long can't be stored as an integer.
Instead it becomes a float and when converted into string it is 6.3041E+15 instead of 6304100000000008.

CreditCard::validate should only allow string arguments.

This is done in a BC way and should be updated for next major.